Send self :connection_ready on boot
It looks like hedwig_flowdock was expecting to handle connection_ready, but never receiving it. I’m not sure if Hedwig used to send this message, but it doesn’t anymore. And in Hedwig’s own Console and Test adapters, it expects the adapter to send itself this message, so I do likewise in the Flowdock adapter here.